requirements.txt
pynxm
discord.py
discordhealthcheck
